Create add-campaign-support-custom-module.adoc
Just creating a relationship between your custom module contacts and Target Lists does not allow you to use the campaign feature like you’re used to with Leads, Targets and Contacts. In order to add campaign functionality to a custom contact module, you must follow the steps in this blog to add the functionality.