Update Config.adoc
The documentation says the config.php files are in the root SuiteCRM folder. But there is no mentioning that the root SuitCRM folder is currently the legacy folder. This is counter intuitive and should be mentioned in some way in the documentation. Therefore i have made a note in brackets of the file path. Other suggestions welcome.